EFS performance boost
The EFS performance is saturated by the redis-check-aof command on
the nexflow prod account.  The result is that it is taking 2hrs
to execute the redis check. The EFS volume contains a few files one
of them being the appendonly.aof file that's very large (~15GB).  We
attempt to boost the EFS performance by configuring higher throughput
for the EFS volume.  Hopefully this will allow redis-check-aof to scan
the appendonly.aof file faster.