Fix nextflow Elasticache setup
Fixed the following failures..

* ElasticacheCluster AWS::ElastiCache::CacheCluster CREATE_FAILED Encryption feature is not supported for engine REDIS.
* nextflow-elasticache-cluster AWS::CloudFormation::Stack ROLLBACK_IN_PROGRESS Template error: The
  ConfigurationEndpoint.Port attribute type is not supported for the redis cache cluster engine.
* nextflow-elasticache-cluster AWS::CloudFormation::Stack ROLLBACK_IN_PROGRESS Template error: The
  ConfigurationEndpoint.Address attribute type is not supported for the redis cache cluster engine.

I was not able to enable transit encryption for Elasticache with Redis.
The AWS docs[1] says it supported but I keep getting errors on multiple
attempts.  We just leave it unencrypted for now to get things working.
We will attempt to encrypt later on.  I do know that encryption is only
supported on r6g instances and not r6gd instances so we are gonna switch
to use the more common instance.

[1] https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/in-transit-encryption.html