[IT-3929] Update MySQL DB cluster
Attempts to do in-place MySql DB upgrades does not work.
The idea here is to do a blue/green upgrade. We will create
a parallel MySql v8.0 DB cluster, migrate the existing
data from the old DB to the new DB then switch the Seqera
app servers to use the new DB.