[IT-3888] Add a test user for py-orca
The py-orca runs tests by trigging an example-dev-project workflow.
We need to add our service user to seqera tower so it will
have access to the tower project.

Note: This user is not in AWS, it's only in seqera tower. We need
to add it as an AWS ARN because the configure-tower-project.py script
takes users from the S3ReadWriteAccessArns and automatically
adds them to the seqera project.