[IT-2361] Update Seqera enterprise version (#477)
This is a 2nd attempt at updating to v25.2.  The first attempt
in #474 failed to deploy and was reverted in #476. Seqera
support says the fix is to update to the latest v25.2 version.

Update to Seqera enterprise to v25.2.3[1]. This is an enterprise
version that supports the AWS cloud compute environment[2]
which would allow getting past the 50 job queue per account
per region constraint.

[1] https://docs.seqera.io/changelog/seqera-enterprise/v25.2.3
[2] https://docs.seqera.io/platform-enterprise/compute-envs/aws-cloud